Date converter util method
